[FLINK-19747][orc] Move Predicate classes to OrcFilters
