+++ flink-formats/flink-orc/src/test/java/org/apache/flink/orc/OrcFileSystemFilterTest.java
@@ -51,20 +51,20 @@ public class OrcFileSystemFilterTest {
 		args.add(valueLiteralExpression);
 
 		CallExpression equalExpression = new CallExpression(BuiltInFunctionDefinitions.EQUALS, args, DataTypes.BOOLEAN());
-		OrcSplitReader.Predicate predicate1 = OrcFilters.toOrcPredicate(equalExpression);
-		OrcSplitReader.Predicate predicate2 = new OrcSplitReader.Equals("long1", PredicateLeaf.Type.LONG, 10);
+		OrcFilters.Predicate predicate1 = OrcFilters.toOrcPredicate(equalExpression);
+		OrcFilters.Predicate predicate2 = new OrcFilters.Equals("long1", PredicateLeaf.Type.LONG, 10);
 		assertTrue(predicate1.toString().equals(predicate2.toString()));
 
 		// greater than
 		CallExpression greaterExpression = new CallExpression(BuiltInFunctionDefinitions.GREATER_THAN, args, DataTypes.BOOLEAN());
-		OrcSplitReader.Predicate predicate3 = OrcFilters.toOrcPredicate(greaterExpression);
-		OrcSplitReader.Predicate predicate4 = new OrcSplitReader.Not(new OrcSplitReader.LessThanEquals("long1", PredicateLeaf.Type.LONG, 10));
+		OrcFilters.Predicate predicate3 = OrcFilters.toOrcPredicate(greaterExpression);
+		OrcFilters.Predicate predicate4 = new OrcFilters.Not(new OrcFilters.LessThanEquals("long1", PredicateLeaf.Type.LONG, 10));
 		assertTrue(predicate3.toString().equals(predicate4.toString()));
 
 		// less than
 		CallExpression lessExpression = new CallExpression(BuiltInFunctionDefinitions.LESS_THAN, args, DataTypes.BOOLEAN());
-		OrcSplitReader.Predicate predicate5 = OrcFilters.toOrcPredicate(lessExpression);
-		OrcSplitReader.Predicate predicate6 = new OrcSplitReader.LessThan("long1", PredicateLeaf.Type.LONG, 10);
+		OrcFilters.Predicate predicate5 = OrcFilters.toOrcPredicate(lessExpression);
+		OrcFilters.Predicate predicate6 = new OrcFilters.LessThan("long1", PredicateLeaf.Type.LONG, 10);
 		assertTrue(predicate5.toString().equals(predicate6.toString()));
 	}
 }
