+++ flink-formats/flink-orc/src/test/java/org/apache/flink/orc/OrcTableSourceTest.java
@@ -221,13 +221,13 @@ public class OrcTableSourceTest {
 		when(environment.createInput(any(InputFormat.class))).thenReturn(mock(DataSource.class));
 		spyTS.getDataSet(environment);
 
-		ArgumentCaptor<OrcSplitReader.Predicate> arguments = ArgumentCaptor.forClass(OrcSplitReader.Predicate.class);
+		ArgumentCaptor<OrcFilters.Predicate> arguments = ArgumentCaptor.forClass(OrcFilters.Predicate.class);
 		verify(mockIF, times(2)).addPredicate(arguments.capture());
 		List<String> values = arguments.getAllValues().stream().map(Object::toString).collect(Collectors.toList());
 		assertTrue(values.contains(
-			new OrcSplitReader.Not(new OrcSplitReader.LessThanEquals("int1", PredicateLeaf.Type.LONG, 100)).toString()));
+			new OrcFilters.Not(new OrcFilters.LessThanEquals("int1", PredicateLeaf.Type.LONG, 100)).toString()));
 		assertTrue(values.contains(
-			new OrcSplitReader.Equals("string1", PredicateLeaf.Type.STRING, "hello").toString()));
+			new OrcFilters.Equals("string1", PredicateLeaf.Type.STRING, "hello").toString()));
 
 		// ensure filter pushdown is correct
 		assertTrue(spyTS.isFilterPushedDown());
